backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
pgvector
python-dotenv
PyMuPDF
python-docx
httpx
email-validator
google-generativeai